requirements.txt
tqdm
timm==0.5.4
glob2==0.7
torchmetrics==0.3.2
albumentations==1.0.0
pytorch_toolbelt==0.4.3
pytorch-lightning==1.3.7.post0
pytorch-metric-learning